Reflection: Support List and MutableList
Lists and MutableLists are supported with the Reflection version of
DeepPrint, as standalone objects or within Data Classes. However, at
runtime there is no distinction between List and MutableList, as they
both are using Java ArrayList. As a standalone object, MutableLists can
be printed as MutableLists, and Lists as Lists, via their extension
functions. If they are parameters, however, they will always print the
mutableListOf() constructor.