Reflection: Map and MutableMap (#21)
Issue #20 

[Issue]
The reflection implementation supports `List` and `MutableList`. We need to do the same for `Map` and `MutableMap`.

[Fix]
Add public extension functions to `Map` and `MutableMap`, and let `deepPrintReflection()` delegate to them.
As with the `List` and `MutableList` implementations, at runtime, there is no difference between a `Map` and a `MutableMap`; we can only get that distinction at compile time, as in the KSP implementation. When the `Map` or `MutableMap` is a standalone object, we can call the appropriate extension function, `deepPrintMapReflection()` or `deepPrintMutableMapReflection(). When it is a property of a `data class`, we use the `mutableMapOf()` constructor to fit both cases.

[Test]
- Unit tests for standalone maps with primitives and objects
- Unit tests for maps as properties of data classes